externalize ALL reporting logic into single file
